Add timeout in RemoteConnection send && fix bugs (#194)
* Fix initializing bug

* Add timeout in RemoteConnection send
