+++ src/main/java/com/webank/wecross/config/P2PServiceConfig.java
@@ -17,21 +17,6 @@ public class P2PServiceConfig {
     @Resource SeqMapper seqMapper;
 
     @Bean
-    public ThreadPoolTaskExecutor newThreadPoolTaskExecutor() {
-        System.out.println("Initializing ThreadPoolTaskExecutor ...");
-        // init default thread pool
-
-        final int threadNum = 8;
-        ThreadPoolTaskExecutor threadPool = new ThreadPoolTaskExecutor();
-        threadPool.setCorePoolSize(threadNum);
-        threadPool.setMaxPoolSize(threadNum);
-        threadPool.setQueueCapacity(1000);
-        threadPool.setThreadNamePrefix("netty-p2p");
-        threadPool.initialize();
-        return threadPool;
-    }
-
-    @Bean
     public P2PService newP2PService() {
         System.out.println("Initializing P2PService ...");
